+++ util/nilawaytest/testhelper.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package testhelper implements utility functions for tests.
-package testhelper
+// Package nilawaytest implements utility functions for tests.
+package nilawaytest
 
 import (
 	"go/ast"
